Use modern upscaling algorithms to improve the cover
The initial cover was high(ish) res but very muddy. The followup used a low-res image to provide colour information and the higher res image for structure. But upscaling algorithms are now good enough to the point that we get a better result by just upscaling the low-res image and ignoring the high-res version completely.